% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof_tests.R
\name{ad_test}
\alias{ad_test}
\title{Teste de Anderson-Darling}
\usage{
ad_test(x, parameters)
}
\arguments{
\item{x}{vetor numérico}

\item{parameters}{parâmetros da distribuição.}
}
\value{
Estatística e p-valor do teste de Anderson-Darling.
}
\description{
Teste de Anderson-Darling
}
