% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof_tests.R
\name{chi_square_test}
\alias{chi_square_test}
\title{Teste de Qui-quadrado}
\usage{
chi_square_test(x, parameters, formula = "weibull")
}
\arguments{
\item{x}{vetor numérico}

\item{parameters}{parâmetros da distribuição.}

\item{formula}{fórmula de plotagem.}
}
\value{
Estatística e p-valor do teste de Qui-quadrado.
}
\description{
Teste de Qui-quadrado
}
