% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_funs.R
\name{density_function}
\alias{density_function}
\title{Ajusta os parâmetros de uma distribuição de probabilidade}
\usage{
density_function(x, parameters)
}
\arguments{
\item{x}{vetor numérico.}

\item{parameters}{parâmetros da distribuição.}
}
\value{
Função densidade de probabilidade.
}
\description{
Ajusta os parâmetros de uma distribuição de probabilidade
}
\examples{
x <- rnorm(50)
par <- fit(x, "nor", "mm")
density_function(x, par)
}
