% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof_tests.R
\name{filliben_test}
\alias{filliben_test}
\title{Teste de Filliben}
\usage{
filliben_test(x, parameters, formula = "weibull")
}
\arguments{
\item{x}{vetor numérico}

\item{parameters}{parâmetros da distribuição.}
}
\value{
Estatística e p-valor do teste de Filliben.
}
\description{
Teste de Filliben
}
