% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof_tests.R
\name{ks_test}
\alias{ks_test}
\title{Teste de Kolmogorov–Smirnov}
\usage{
ks_test(x, parameters, formula = "weibull")
}
\arguments{
\item{x}{vetor numérico}

\item{parameters}{parâmetros da distribuição.}

\item{formula}{fórmula de plotagem.}
}
\value{
Estatística e p-valor do teste de Kolmogorov–Smirnov.
}
\description{
Teste de Kolmogorov–Smirnov
}
