% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_funs.R
\name{print.density_function}
\alias{print.density_function}
\title{print method}
\usage{
\method{print}{density_function}(x, ...)
}
\arguments{
\item{x}{S3 object of class density_function.}
}
\description{
print method
}
