% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_formulas.R
\name{print.plotting_position}
\alias{print.plotting_position}
\title{print method}
\usage{
\method{print}{plotting_position}(x, ...)
}
\arguments{
\item{x}{S3 object of class q_function.}
}
\description{
print method
}
