% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_funs.R
\name{prob_function}
\alias{prob_function}
\title{Ajusta os parâmetros de uma distribuição de probabilidade}
\usage{
prob_function(x, parameters)
}
\arguments{
\item{x}{vetor numérico.}

\item{parameters}{parâmetros da distribuição.}
}
\value{
Função cumulativa de probabilidade.
}
\description{
Ajusta os parâmetros de uma distribuição de probabilidade
}
\examples{
x <- rnorm(50)
par <- fit(x, "nor", "mm")
prob_function(x, par)
}
