% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_funs.R
\name{q_function}
\alias{q_function}
\title{Função quantil}
\usage{
q_function(x, parameters)
}
\arguments{
\item{x}{vetor de probabilidades.}

\item{parameters}{parâmetros da distribuição.}
}
\value{
Quantis.
}
\description{
Função quantil
}
\examples{
x <- rnorm(50)
par <- fit(x, "nor", "mm")
q_function(x, par)
}
