% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_formulas.R
\name{plotting_position}
\alias{plotting_position}
\title{Posição de plotagem}
\usage{
plotting_position(x, formula = "weibull", exced = FALSE)
}
\arguments{
\item{x}{vetor numérico.}

\item{formula}{weibull, gringorten, blom, hazem ou cunnane.}

\item{exced}{logical; probabilidade de excedência.}
}
\value{
posições de plotagem para a fórmula selecionada.
}
\description{
Posição de plotagem
}
\examples{
x <- rnorm(50)
plotting_position(x, formula = "weibull")
}
