% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_funs.R
\name{prob_function}
\alias{prob_function}
\title{Funçãode distribuição de probabilidade}
\usage{
prob_function(x, parameters, exced = FALSE)
}
\arguments{
\item{x}{vetor numérico.}

\item{parameters}{parâmetros da distribuição.}

\item{exced}{logical; probabilidade de excedência.}
}
\value{
Função cumulativa de probabilidade.
}
\description{
Funçãode distribuição de probabilidade
}
\examples{
x <- rnorm(50)
par <- fit(x, "nor", "mm")
prob_function(x, par)
}
